// frontend/aada_web/src/api/generated/models/refundRead.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * AADA LMS API
 * OpenAPI spec version: 1.0
 */
import type { RefundReadPolicyBasis } from './refundReadPolicyBasis';
import type { RefundReadApprovedBy } from './refundReadApprovedBy';
import type { RefundReadApprovedAt } from './refundReadApprovedAt';
import type { RefundReadRemittedAt } from './refundReadRemittedAt';

export interface RefundRead {
  withdrawal_id: string;
  policy_basis?: RefundReadPolicyBasis;
  approved_by?: RefundReadApprovedBy;
  approved_at?: RefundReadApprovedAt;
  remitted_at?: RefundReadRemittedAt;
  id: string;
  amount_cents: number;
}