// frontend/aada_web/src/api/generated/models/withdrawalRead.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * AADA LMS API
 * OpenAPI spec version: 1.0
 */
import type { WithdrawalReadReason } from './withdrawalReadReason';
import type { WithdrawalReadProgressPctAtWithdrawal } from './withdrawalReadProgressPctAtWithdrawal';
import type { WithdrawalReadAdminProcessedAt } from './withdrawalReadAdminProcessedAt';

export interface WithdrawalRead {
  enrollment_id: string;
  requested_at: string;
  reason?: WithdrawalReadReason;
  progress_pct_at_withdrawal?: WithdrawalReadProgressPctAtWithdrawal;
  id: string;
  admin_processed_at?: WithdrawalReadAdminProcessedAt;
}